CI: Test Python 3.11 on all OS, 3.10 and lower on Ubuntu (#423)
Updates the job matrix of the GitHub Actions test workflow to test with Python 3.11 on all operating systems. 3.8, 3.9 and 3.10 are tested with one OS, Ubuntu. This configurations keeps the number of jobs reasonable at 6, while testing all operating systems and four Python versions.

Also sets fail-fast to false, so that other runs continue if one fails.