Filter column indices of sparse matrix to ensure non-negativity
In Issue #573, I observed that `scipy`s compressed sparse column matrix
was throwing an exception about specification of negative indices. I
traced this issue to the column specification in the C++ implementation
of RMTB.

I could not figure out what precisely was going wrong. However, by
simply replacing all negative column indices with zeros, I found that
the unit tests pass. This is in no way a permanent solution, but given
the lack of shared understanding of this code, it is preferable to it
being fully nonfunctional.

While we're at it, replace all internal `int`s by `long` in order to
support larger model sizes, similar to the diff made in RMTC.