Document example with less expensive parameters
The documentation for RMTC uses `num_elements=20`.
In Issue #574, it is shown that the training time is exponential in
`num_elements`.
Hence using copy-paste from the example can blow up the compute time.

Reduce `num_elements` in the example so that users are less likely to
hit this exponential behavior.