New test for squar_sin_exp
new definition of a problem for squar_sin_exp,
as the previous one was modeled by a constant line in one direction,
creating a division by zero in the computation of the relative error.
Also, a new computation for the error on the derivative of the variance
is introduced, in order to account for when value can be so close
to zero that the relative difference doesn't make much sense anymore.